Refactored a lot.  Moved stuff around to cleanup repetitive code.
