+++ CHRISUpdate/Utilities/CSVHelper.cs
@@ -1,0 +1,67 @@
+﻿using CsvHelper;
+using CsvHelper.Configuration;
+using System;
+using System.Collections.Generic;
+using System.Configuration;
+using System.IO;
+using System.Linq;
+
+namespace HRUpdate.Utilities
+{
+    internal class FileReader
+    {
+        public FileReader()
+        {
+        }
+
+        public List<TClass> GetFileData<TClass, TMap>(string filePath)
+            where TClass : class
+            where TMap : ClassMap<TClass>
+        {
+            CsvParser csvParser = new CsvParser(new StreamReader(filePath));
+            CsvReader csvReader = new CsvReader(csvParser);
+
+            csvReader.Configuration.Delimiter = "~";
+            csvReader.Configuration.HasHeaderRecord = false;
+
+            csvReader.Configuration.RegisterClassMap<TMap>();
+
+            return csvReader.GetRecords<TClass>().ToList();
+        }
+    }
+
+    internal class SummaryFileGenerator
+    {
+        //Reference to logger
+        private static readonly log4net.ILog log = log4net.LogManager.GetLogger(System.Reflection.MethodBase.GetCurrentMethod().DeclaringType);
+
+        //Empty Constructor
+        public SummaryFileGenerator() { }
+
+        internal string GenerateSummaryFile<TClass, TMap>(string fileName, List<TClass> summaryData)
+            where TClass : class
+            where TMap : ClassMap<TClass>
+        {
+            try
+            {
+                string summaryFileName;
+
+                summaryFileName = fileName + "_" + DateTime.Now.ToString("yyyyMMddHHmmss_FFFF") + ".csv";
+
+                //Creates the summary file
+                using (CsvWriter csvWriter = new CsvWriter(new StreamWriter(ConfigurationManager.AppSettings["SUMMARYFILEPATH"] + summaryFileName, false)))
+                {
+                    csvWriter.Configuration.RegisterClassMap<TMap>();
+                    csvWriter.WriteRecords(summaryData);
+                }
+
+                return summaryFileName;
+            }
+            catch (Exception ex)
+            {
+                log.Error("Error Writing Summary File: " + fileName + " - " + ex.Message + " - " + ex.InnerException);
+                return string.Empty;
+            }
+        }
+    }
+}
